+++ waterfall-cad/src/Waterfall/Sweep.hs
@@ -4,8 +4,8 @@ module Waterfall.Sweep
 
 import Waterfall.Internal.Solid (Solid (..))
 import Waterfall.Internal.Path (Path (..))
-import Waterfall.Internal.Edges (wireTangent)
-import Waterfall.Transforms (rotate)
+import Waterfall.Internal.Edges (wireTangent, wireEndpoints)
+import Waterfall.Transforms (rotate, translate)
 import Waterfall.TwoD.Internal.Shape (Shape (..))
 import qualified OpenCascade.BRepOffsetAPI.MakePipe as MakePipe
 import qualified OpenCascade.BRepBuilderAPI.MakeShape as MakeShape
@@ -16,8 +16,8 @@ import Foreign.Ptr
 import Linear (V3, normalize, unit, _x, _z, nearZero, cross, dot)
 import Data.Acquire (Acquire)
 
-adjustFace :: V3 Double -> Ptr TopoDS.Shape -> Acquire (Ptr TopoDS.Shape)
-adjustFace v face = 
+rotateFace :: V3 Double -> Ptr TopoDS.Shape -> Acquire (Ptr TopoDS.Shape)
+rotateFace v face = 
     let vn = normalize v
         z = unit _z
         in if nearZero (vn - z)
@@ -26,6 +26,9 @@ adjustFace v face =
                 let axis = if nearZero (vn + z) then unit _x else cross vn z
                     angle = acos (dot vn z) 
                 in runSolid . rotate axis angle . Solid . pure $ face 
+
+positionFace :: V3 Double -> Ptr TopoDS.Shape -> Acquire (Ptr TopoDS.Shape)
+positionFace p = runSolid . translate p . Solid . pure
 
 -- | Sweep a 2D `Shape` along a `Path`, constructing a `Solid`
 -- 
@@ -37,6 +40,7 @@ sweep (Path runThePath) (Shape runTheShape) = Solid $ do
     path <- runThePath
     shape <- runTheShape
     tangent <- liftIO $ wireTangent path
-    adjustedFace <- adjustFace tangent shape
+    (start,_)  <- liftIO $ wireEndpoints path
+    adjustedFace <- positionFace start =<< rotateFace tangent shape
     builder <- MakePipe.fromWireAndShape path adjustedFace
     MakeShape.shape (upcast builder)